refactor según especificaciones corrección. Separación de controladores y vistas, solo print console en vistas
