+++ src/Main.java
@@ -1,14 +1,18 @@
-import datarsians.controlador.Controlador;
+
+import datarsians.controlador.ControladorArticulo;
+import datarsians.controlador.ControladorCliente;
+import datarsians.controlador.ControladorPedidos;
 import datarsians.modelo.Datos;
 import datarsians.vista.Menu;
 
-public class Main {
-    public static void main(String[] args) {
-        Datos datos = new Datos();
-        // datos.cargarDatosDePrueba();
+public static void main(String[] args) {
+    Datos datos = new Datos();
+    // System.out.println( datos.cargarDatosDePrueba() );
 
-        Controlador controlador = new Controlador(datos);
-        Menu menu = new Menu(controlador);
-        menu.mostrarMenuPrincipal();
-    }
+    Menu menu = new Menu(
+            new ControladorArticulo(datos),
+            new ControladorCliente(datos),
+            new ControladorPedidos(datos)
+    );
+    menu.mostrarMenuPrincipal();
 }
