--- src/datarsians/vista/prueba.java
@@ -1,4 +1,0 @@
-package datarsians.vista;
-
-public class prueba {
-}
